pyproject.toml
[tool.poetry]
name = "yakrazor"
version = "0.1.0"
description = "Yak Razor: Scheduler for Yak Shavers"
authors = ["Harshad Sharma <harshad@sharma.io>"]
readme = "README.md"
packages = [{include = "yakrazor", from = "src"}]

[tool.poetry.scripts]
yakrazor = "yakrazor.app:run_app"

[tool.poetry.dependencies]
python = "^3.9"
tortoise-orm = "^0.20.0"
nicegui = "^1.4.18"
appdirs = "^1.4.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
